+++ ShapesJava/src/org/tyaa/shapesjava/IScalable.java
@@ -2,5 +2,5 @@ package org.tyaa.shapesjava;
 
 public interface IScalable {
 
-    public void scale(Integer _zoom);
+    public void scale(Double _zoom);
 }
